third_party/blink/web_tests/external/wpt/css/css-shadow/font-face-007-expected.txt

This is a testharness.js-based test.
[FAIL] @font-face from shadow applies to :host
  assert_equals: expected 160 but got 80
Harness: the test ran to completion.
